moved over simplified trans passes to file, copied for making real versions
